schema: refactor license
