+++ .env
@@ -1,4 +1,4 @@
-SERLO_ORG_HOST=serlo.localhost:4567
+SERLO_ORG_HOST="serlo.org"
 SERLO_ORG_SECRET="serlo.org-secret"
 PLAYGROUND_SECRET="api.serlo.org-playground-secret"
 HYDRA_HOST="http://127.0.0.1:4445"
