+++ .eslintrc.js
@@ -70,7 +70,12 @@ module.exports = {
         optionalDependencies: false,
       },
     ],
-    'import/no-internal-modules': 'error',
+    'import/no-internal-modules': [
+      'error',
+      {
+        allow: ['msw/node'],
+      },
+    ],
     'import/no-mutable-exports': 'error',
     'import/no-self-import': 'error',
     'import/no-unassigned-import': 'error',
