+++ __tests-pacts__/serlo.org/uuid/applet.ts
@@ -1,39 +1,52 @@
+/**
+ * This file is part of Serlo.org API
+ *
+ * Copyright (c) 2020 Serlo Education e.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License")
+ * you may not use this file except in compliance with the License
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * @copyright Copyright (c) 2020 Serlo Education e.V.
+ * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
+ * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
+ */
 import { gql } from 'apollo-server'
 import * as R from 'ramda'
 
-import { license } from '../../__fixtures__/license'
+import { license } from '../../../__fixtures__/license'
 import {
-  article,
-  articleAlias,
-  articleRevision,
+  applet,
+  appletRevision,
   taxonomyTermSubject,
   user,
-} from '../../__fixtures__/uuid'
-import { assertSuccessfulGraphQLQuery } from '../__utils__/assertions'
+} from '../../../__fixtures__/uuid'
+import { assertSuccessfulGraphQLQuery } from '../../__utils__/assertions'
 import {
-  addAliasInteraction,
-  addArticleInteraction,
-  addArticleRevisionInteraction,
+  addAppletInteraction,
+  addAppletRevisionInteraction,
   addLicenseInteraction,
   addTaxonomyTermInteraction,
   addUserInteraction,
-} from '../__utils__/interactions'
+} from '../../__utils__/interactions'
 
-describe('Article', () => {
-  test('by alias', async () => {
-    await addArticleInteraction(article)
-    await addAliasInteraction(articleAlias)
+describe('Applet', () => {
+  test('by id', async () => {
+    await addAppletInteraction(applet)
     await assertSuccessfulGraphQLQuery({
       query: gql`
           {
-            uuid(
-              alias: {
-                instance: de
-                path: "${articleAlias.path}"
-              }
-            ) {
+            uuid(id: ${applet.id}) {
               __typename
-              ... on Article {
+              ... on Applet {
                 id
                 trashed
                 instance
@@ -51,13 +64,13 @@ describe('Article', () => {
         `,
       data: {
         uuid: {
-          __typename: 'Article',
+          __typename: 'Applet',
           ...R.omit(
             ['currentRevisionId', 'licenseId', 'taxonomyTermIds'],
-            article
+            applet
           ),
           currentRevision: {
-            id: 30674,
+            id: appletRevision.id,
           },
           license: {
             id: 1,
@@ -67,21 +80,15 @@ describe('Article', () => {
     })
   })
 
-  test('by alias (w/ license)', async () => {
-    await addArticleInteraction(article)
-    await addAliasInteraction(articleAlias)
+  test('by id (w/ license)', async () => {
+    await addAppletInteraction(applet)
     await addLicenseInteraction(license)
     await assertSuccessfulGraphQLQuery({
       query: gql`
         {
-          uuid(
-            alias: {
-              instance: de
-              path: "${articleAlias.path}"
-            }
-          ) {
+          uuid(id: ${applet.id}) {
             __typename
-            ... on Article {
+            ... on Applet {
               id
               trashed
               instance
@@ -100,13 +107,13 @@ describe('Article', () => {
       `,
       data: {
         uuid: {
-          __typename: 'Article',
+          __typename: 'Applet',
           ...R.omit(
             ['currentRevisionId', 'licenseId', 'taxonomyTermIds'],
-            article
+            applet
           ),
           currentRevision: {
-            id: 30674,
+            id: appletRevision.id,
           },
           license: {
             id: 1,
@@ -117,21 +124,15 @@ describe('Article', () => {
     })
   })
 
-  test('by alias (w/ currentRevision)', async () => {
-    await addArticleInteraction(article)
-    await addAliasInteraction(articleAlias)
-    await addArticleRevisionInteraction(articleRevision)
+  test('by id (w/ currentRevision)', async () => {
+    await addAppletInteraction(applet)
+    await addAppletRevisionInteraction(appletRevision)
     await assertSuccessfulGraphQLQuery({
       query: gql`
           {
-            uuid(
-              alias: {
-                instance: de
-                path: "${articleAlias.path}"
-              }
-            ) {
+            uuid(id: ${applet.id}) {
               __typename
-              ... on Article {
+              ... on Applet {
                 id
                 trashed
                 instance
@@ -149,13 +150,13 @@ describe('Article', () => {
         `,
       data: {
         uuid: {
-          __typename: 'Article',
+          __typename: 'Applet',
           ...R.omit(
             ['currentRevisionId', 'licenseId', 'taxonomyTermIds'],
-            article
+            applet
           ),
           currentRevision: {
-            id: 30674,
+            id: appletRevision.id,
             title: 'title',
             content: 'content',
             changes: 'changes',
@@ -165,21 +166,15 @@ describe('Article', () => {
     })
   })
 
-  test('by alias (w/ taxonomyTerms)', async () => {
-    await addArticleInteraction(article)
-    await addAliasInteraction(articleAlias)
+  test('by id (w/ taxonomyTerms)', async () => {
+    await addAppletInteraction(applet)
     await addTaxonomyTermInteraction(taxonomyTermSubject)
     await assertSuccessfulGraphQLQuery({
       query: gql`
         {
-          uuid(
-            alias: {
-              instance: de
-              path: "/mathe/funktionen/uebersicht-aller-artikel-zu-funktionen/parabel"
-            }
-          ) {
+          uuid(id: ${applet.id}) {
             __typename
-            ... on Article {
+            ... on Applet {
               id
               trashed
               instance
@@ -194,80 +189,40 @@ describe('Article', () => {
       `,
       data: {
         uuid: {
-          __typename: 'Article',
+          __typename: 'Applet',
           ...R.omit(
             ['currentRevisionId', 'licenseId', 'taxonomyTermIds'],
-            article
+            applet
           ),
           taxonomyTerms: [{ id: 5 }],
         },
       },
     })
   })
-
-  test('by id', async () => {
-    await addArticleInteraction(article)
-    await assertSuccessfulGraphQLQuery({
-      query: gql`
-        {
-          uuid(id: 1855) {
-            __typename
-            ... on Article {
-              id
-              trashed
-              alias
-              instance
-              date
-              currentRevision {
-                id
-              }
-              license {
-                id
-              }
-            }
-          }
-        }
-      `,
-      data: {
-        uuid: {
-          __typename: 'Article',
-          ...R.omit(
-            ['currentRevisionId', 'licenseId', 'taxonomyTermIds'],
-            article
-          ),
-          currentRevision: {
-            id: 30674,
-          },
-          license: {
-            id: 1,
-          },
-        },
-      },
-    })
-  })
 })
 
-describe('ArticleRevision', () => {
+describe('AppletRevision', () => {
   test('by id', async () => {
-    await addArticleRevisionInteraction(articleRevision)
+    await addAppletRevisionInteraction(appletRevision)
     await assertSuccessfulGraphQLQuery({
       query: gql`
         {
-          uuid(id: 30674) {
+          uuid(id: ${appletRevision.id}) {
             __typename
-            ... on ArticleRevision {
+            ... on AppletRevision {
               id
               trashed
               date
               title
               content
+              url
               changes
               metaTitle
               metaDescription
               author {
                 id
               }
-              article {
+              applet {
                 id
               }
             }
@@ -276,13 +231,13 @@ describe('ArticleRevision', () => {
       `,
       data: {
         uuid: {
-          __typename: 'ArticleRevision',
-          ...R.omit(['authorId', 'repositoryId'], articleRevision),
+          __typename: 'AppletRevision',
+          ...R.omit(['authorId', 'repositoryId'], appletRevision),
           author: {
             id: 1,
           },
-          article: {
-            id: 1855,
+          applet: {
+            id: applet.id,
           },
         },
       },
@@ -290,19 +245,20 @@ describe('ArticleRevision', () => {
   })
 
   test('by id (w/ author)', async () => {
-    await addArticleRevisionInteraction(articleRevision)
+    await addAppletRevisionInteraction(appletRevision)
     await addUserInteraction(user)
     await assertSuccessfulGraphQLQuery({
       query: gql`
         {
-          uuid(id: 30674) {
+          uuid(id: ${appletRevision.id}) {
             __typename
-            ... on ArticleRevision {
+            ... on AppletRevision {
               id
               trashed
               date
               title
               content
+              url
               changes
               metaTitle
               metaDescription
@@ -310,7 +266,7 @@ describe('ArticleRevision', () => {
                 id
                 username
               }
-              article {
+              applet {
                 id
               }
             }
@@ -319,41 +275,42 @@ describe('ArticleRevision', () => {
       `,
       data: {
         uuid: {
-          __typename: 'ArticleRevision',
-          ...R.omit(['authorId', 'repositoryId'], articleRevision),
+          __typename: 'AppletRevision',
+          ...R.omit(['authorId', 'repositoryId'], appletRevision),
           author: {
             id: 1,
             username: user.username,
           },
-          article: {
-            id: 1855,
+          applet: {
+            id: applet.id,
           },
         },
       },
     })
   })
 
-  test('by id (w/ article)', async () => {
-    await addArticleRevisionInteraction(articleRevision)
-    await addArticleInteraction(article)
+  test('by id (w/ applet)', async () => {
+    await addAppletRevisionInteraction(appletRevision)
+    await addAppletInteraction(applet)
     await assertSuccessfulGraphQLQuery({
       query: gql`
         {
-          uuid(id: 30674) {
+          uuid(id: ${appletRevision.id}) {
             __typename
-            ... on ArticleRevision {
+            ... on AppletRevision {
               id
               trashed
               date
               title
               content
+              url
               changes
               metaTitle
               metaDescription
               author {
                 id
               }
-              article {
+              applet {
                 id
                 currentRevision {
                   id
@@ -365,15 +322,15 @@ describe('ArticleRevision', () => {
       `,
       data: {
         uuid: {
-          __typename: 'ArticleRevision',
-          ...R.omit(['authorId', 'repositoryId'], articleRevision),
+          __typename: 'AppletRevision',
+          ...R.omit(['authorId', 'repositoryId'], appletRevision),
           author: {
             id: 1,
           },
-          article: {
-            id: 1855,
+          applet: {
+            id: applet.id,
             currentRevision: {
-              id: 30674,
+              id: appletRevision.id,
             },
           },
         },
