+++ __tests-pacts__/serlo.org/uuid/course.ts
@@ -1,39 +1,64 @@
+/**
+ * This file is part of Serlo.org API
+ *
+ * Copyright (c) 2020 Serlo Education e.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License")
+ * you may not use this file except in compliance with the License
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * @copyright Copyright (c) 2020 Serlo Education e.V.
+ * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
+ * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
+ */
 import { gql } from 'apollo-server'
 import * as R from 'ramda'
 
-import { license } from '../../__fixtures__/license'
+import { license } from '../../../__fixtures__/license'
 import {
-  event,
-  eventAlias,
-  eventRevision,
+  course,
+  courseAlias,
+  coursePage,
+  coursePageRevision,
+  courseRevision,
   taxonomyTermSubject,
   user,
-} from '../../__fixtures__/uuid'
-import { assertSuccessfulGraphQLQuery } from '../__utils__/assertions'
+} from '../../../__fixtures__/uuid'
+import { assertSuccessfulGraphQLQuery } from '../../__utils__/assertions'
 import {
   addAliasInteraction,
-  addEventInteraction,
-  addEventRevisionInteraction,
+  addCourseInteraction,
+  addCoursePageInteraction,
+  addCoursePageRevisionInteraction,
+  addCourseRevisionInteraction,
   addLicenseInteraction,
   addTaxonomyTermInteraction,
   addUserInteraction,
-} from '../__utils__/interactions'
+} from '../../__utils__/interactions'
 
-describe('Event', () => {
+describe('Course', () => {
   test('by alias', async () => {
-    await addEventInteraction(event)
-    await addAliasInteraction(eventAlias)
+    await addCourseInteraction(course)
+    await addAliasInteraction(courseAlias)
     await assertSuccessfulGraphQLQuery({
       query: gql`
           {
             uuid(
               alias: {
                 instance: de
-                path: "${eventAlias.path}"
+                path: "${courseAlias.path}"
               }
             ) {
               __typename
-              ... on Event {
+              ... on Course {
                 id
                 trashed
                 instance
@@ -51,13 +76,13 @@ describe('Event', () => {
         `,
       data: {
         uuid: {
-          __typename: 'Event',
+          __typename: 'Course',
           ...R.omit(
-            ['currentRevisionId', 'licenseId', 'taxonomyTermIds'],
-            event
+            ['currentRevisionId', 'licenseId', 'taxonomyTermIds', 'pageIds'],
+            course
           ),
           currentRevision: {
-            id: eventRevision.id,
+            id: course.currentRevisionId,
           },
           license: {
             id: 1,
@@ -68,8 +93,8 @@ describe('Event', () => {
   })
 
   test('by alias (w/ license)', async () => {
-    await addEventInteraction(event)
-    await addAliasInteraction(eventAlias)
+    await addCourseInteraction(course)
+    await addAliasInteraction(courseAlias)
     await addLicenseInteraction(license)
     await assertSuccessfulGraphQLQuery({
       query: gql`
@@ -77,11 +102,11 @@ describe('Event', () => {
           uuid(
             alias: {
               instance: de
-              path: "${eventAlias.path}"
+              path: "${courseAlias.path}"
             }
           ) {
             __typename
-            ... on Event {
+            ... on Course {
               id
               trashed
               instance
@@ -100,13 +125,13 @@ describe('Event', () => {
       `,
       data: {
         uuid: {
-          __typename: 'Event',
+          __typename: 'Course',
           ...R.omit(
-            ['currentRevisionId', 'licenseId', 'taxonomyTermIds'],
-            event
+            ['currentRevisionId', 'licenseId', 'taxonomyTermIds', 'pageIds'],
+            course
           ),
           currentRevision: {
-            id: eventRevision.id,
+            id: course.currentRevisionId,
           },
           license: {
             id: 1,
@@ -118,20 +143,20 @@ describe('Event', () => {
   })
 
   test('by alias (w/ currentRevision)', async () => {
-    await addEventInteraction(event)
-    await addAliasInteraction(eventAlias)
-    await addEventRevisionInteraction(eventRevision)
+    await addCourseInteraction(course)
+    await addAliasInteraction(courseAlias)
+    await addCourseRevisionInteraction(courseRevision)
     await assertSuccessfulGraphQLQuery({
       query: gql`
           {
             uuid(
               alias: {
                 instance: de
-                path: "${eventAlias.path}"
+                path: "${courseAlias.path}"
               }
             ) {
               __typename
-              ... on Event {
+              ... on Course {
                 id
                 trashed
                 instance
@@ -139,7 +164,6 @@ describe('Event', () => {
                 date
                 currentRevision {
                   id
-                  title
                   content
                   changes
                 }
@@ -149,14 +173,13 @@ describe('Event', () => {
         `,
       data: {
         uuid: {
-          __typename: 'Event',
+          __typename: 'Course',
           ...R.omit(
-            ['currentRevisionId', 'licenseId', 'taxonomyTermIds'],
-            event
+            ['currentRevisionId', 'licenseId', 'taxonomyTermIds', 'pageIds'],
+            course
           ),
           currentRevision: {
-            id: eventRevision.id,
-            title: 'title',
+            id: course.currentRevisionId,
             content: 'content',
             changes: 'changes',
           },
@@ -166,8 +189,8 @@ describe('Event', () => {
   })
 
   test('by alias (w/ taxonomyTerms)', async () => {
-    await addEventInteraction(event)
-    await addAliasInteraction(eventAlias)
+    await addCourseInteraction(course)
+    await addAliasInteraction(courseAlias)
     await addTaxonomyTermInteraction(taxonomyTermSubject)
     await assertSuccessfulGraphQLQuery({
       query: gql`
@@ -175,11 +198,11 @@ describe('Event', () => {
           uuid(
             alias: {
               instance: de
-              path: "${eventAlias.path}"
+              path: "${courseAlias.path}"
             }
           ) {
             __typename
-            ... on Event {
+            ... on Course {
               id
               trashed
               instance
@@ -194,10 +217,10 @@ describe('Event', () => {
       `,
       data: {
         uuid: {
-          __typename: 'Event',
+          __typename: 'Course',
           ...R.omit(
-            ['currentRevisionId', 'licenseId', 'taxonomyTermIds'],
-            event
+            ['currentRevisionId', 'licenseId', 'taxonomyTermIds', 'pageIds'],
+            course
           ),
           taxonomyTerms: [{ id: 5 }],
         },
@@ -205,14 +228,69 @@ describe('Event', () => {
     })
   })
 
-  test('by id', async () => {
-    await addEventInteraction(event)
+  test('by alias (w/ pages)', async () => {
+    await addCourseInteraction(course)
+    await addAliasInteraction(courseAlias)
+    await addCoursePageRevisionInteraction(coursePageRevision)
+    await addCoursePageInteraction(coursePage)
     await assertSuccessfulGraphQLQuery({
       query: gql`
         {
-          uuid(id: ${event.id}) {
+          uuid(
+            alias: {
+              instance: de
+              path: "${courseAlias.path}"
+            }
+          ) {
             __typename
-            ... on Event {
+            ... on Course {
+              id
+              trashed
+              instance
+              alias
+              date
+              pages {
+                id
+                currentRevision {
+                  id
+                  content
+                  changes
+                }
+              }
+            }
+          }
+        }
+      `,
+      data: {
+        uuid: {
+          __typename: 'Course',
+          ...R.omit(
+            ['currentRevisionId', 'licenseId', 'taxonomyTermIds', 'pageIds'],
+            course
+          ),
+          pages: [
+            {
+              id: coursePage.id,
+              currentRevision: {
+                id: coursePageRevision.id,
+                content: coursePageRevision.content,
+                changes: coursePageRevision.changes,
+              },
+            },
+          ],
+        },
+      },
+    })
+  })
+
+  test('by id', async () => {
+    await addCourseInteraction(course)
+    await assertSuccessfulGraphQLQuery({
+      query: gql`
+        {
+          uuid(id: ${course.id}) {
+            __typename
+            ... on Course {
               id
               trashed
               alias
@@ -230,13 +308,13 @@ describe('Event', () => {
       `,
       data: {
         uuid: {
-          __typename: 'Event',
+          __typename: 'Course',
           ...R.omit(
-            ['currentRevisionId', 'licenseId', 'taxonomyTermIds'],
-            event
+            ['currentRevisionId', 'licenseId', 'taxonomyTermIds', 'pageIds'],
+            course
           ),
           currentRevision: {
-            id: eventRevision.id,
+            id: course.currentRevisionId,
           },
           license: {
             id: 1,
@@ -247,27 +325,26 @@ describe('Event', () => {
   })
 })
 
-describe('EventRevision', () => {
+describe('CourseRevision', () => {
   test('by id', async () => {
-    await addEventRevisionInteraction(eventRevision)
+    await addCourseRevisionInteraction(courseRevision)
     await assertSuccessfulGraphQLQuery({
       query: gql`
         {
-          uuid(id: ${eventRevision.id}) {
+          uuid(id: ${course.currentRevisionId}) {
             __typename
-            ... on EventRevision {
+            ... on CourseRevision {
               id
               trashed
               date
               title
               content
               changes
-              metaTitle
               metaDescription
               author {
                 id
               }
-              event {
+              course {
                 id
               }
             }
@@ -276,13 +353,13 @@ describe('EventRevision', () => {
       `,
       data: {
         uuid: {
-          __typename: 'EventRevision',
-          ...R.omit(['authorId', 'repositoryId'], eventRevision),
+          __typename: 'CourseRevision',
+          ...R.omit(['authorId', 'repositoryId'], courseRevision),
           author: {
             id: 1,
           },
-          event: {
-            id: event.id,
+          course: {
+            id: course.id,
           },
         },
       },
@@ -290,27 +367,26 @@ describe('EventRevision', () => {
   })
 
   test('by id (w/ author)', async () => {
-    await addEventRevisionInteraction(eventRevision)
+    await addCourseRevisionInteraction(courseRevision)
     await addUserInteraction(user)
     await assertSuccessfulGraphQLQuery({
       query: gql`
         {
-          uuid(id: ${eventRevision.id}) {
+          uuid(id: ${course.currentRevisionId}) {
             __typename
-            ... on EventRevision {
+            ... on CourseRevision {
               id
               trashed
               date
               title
               content
               changes
-              metaTitle
               metaDescription
               author {
                 id
                 username
               }
-              event {
+              course {
                 id
               }
             }
@@ -319,41 +395,40 @@ describe('EventRevision', () => {
       `,
       data: {
         uuid: {
-          __typename: 'EventRevision',
-          ...R.omit(['authorId', 'repositoryId'], eventRevision),
+          __typename: 'CourseRevision',
+          ...R.omit(['authorId', 'repositoryId'], courseRevision),
           author: {
             id: 1,
             username: user.username,
           },
-          event: {
-            id: event.id,
+          course: {
+            id: course.id,
           },
         },
       },
     })
   })
 
-  test('by id (w/ event)', async () => {
-    await addEventRevisionInteraction(eventRevision)
-    await addEventInteraction(event)
+  test('by id (w/ course)', async () => {
+    await addCourseRevisionInteraction(courseRevision)
+    await addCourseInteraction(course)
     await assertSuccessfulGraphQLQuery({
       query: gql`
         {
-          uuid(id: ${eventRevision.id}) {
+          uuid(id: ${course.currentRevisionId}) {
             __typename
-            ... on EventRevision {
+            ... on CourseRevision {
               id
               trashed
               date
               title
               content
               changes
-              metaTitle
               metaDescription
               author {
                 id
               }
-              event {
+              course {
                 id
                 currentRevision {
                   id
@@ -365,15 +440,15 @@ describe('EventRevision', () => {
       `,
       data: {
         uuid: {
-          __typename: 'EventRevision',
-          ...R.omit(['authorId', 'repositoryId'], eventRevision),
+          __typename: 'CourseRevision',
+          ...R.omit(['authorId', 'repositoryId'], courseRevision),
           author: {
             id: 1,
           },
-          event: {
-            id: event.id,
+          course: {
+            id: course.id,
             currentRevision: {
-              id: eventRevision.id,
+              id: course.currentRevisionId,
             },
           },
         },
