--- __tests-pacts__/uuid/index.ts
@@ -1,41 +1,0 @@
-/* eslint-disable import/no-unassigned-import */
-describe('Applet', () => {
-  require('./applet')
-})
-describe('Article', () => {
-  require('./article')
-})
-describe('Course', () => {
-  require('./course')
-})
-describe('Course Page', () => {
-  require('./course-page')
-})
-describe('Event', () => {
-  require('./event')
-})
-describe('Exercise', () => {
-  require('./exercise')
-})
-describe('Exercise Group', () => {
-  require('./exercise-group')
-})
-describe('Grouped Exercise', () => {
-  require('./grouped-exercise')
-})
-describe('Page', () => {
-  require('./page')
-})
-describe('Solution', () => {
-  require('./solution')
-})
-describe('Taxonomy Term', () => {
-  require('./taxonomy-term')
-})
-describe('User', () => {
-  require('./user')
-})
-describe('Video', () => {
-  require('./video')
-})
-/* eslint-enable import/no-unassigned-import */
