+++ __tests__/__utils__/assertions.ts
@@ -1,3 +1,24 @@
+/**
+ * This file is part of Serlo.org API
+ *
+ * Copyright (c) 2020 Serlo Education e.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License")
+ * you may not use this file except in compliance with the License
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * @copyright Copyright (c) 2020 Serlo Education e.V.
+ * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
+ * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
+ */
 import { GraphQLResponse } from 'apollo-server-types'
 import { DocumentNode, GraphQLFormattedError } from 'graphql'
 
@@ -5,15 +26,19 @@ import { Client } from './test-client'
 
 export async function assertSuccessfulGraphQLQuery({
   query,
+  variables,
   data,
   client,
 }: {
   query: DocumentNode
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  variables?: Record<string, any>
   data: GraphQLResponse['data']
   client: Client
 }) {
   const response = await client.query({
     query,
+    variables,
   })
   expect(response.errors).toBeUndefined()
   expect(response.data).toEqual(data)
