+++ package.json
@@ -66,6 +66,7 @@
     "eslint-plugin-react": "^7.0.0",
     "glob": "^7.0.0",
     "jest": "^26.0.0",
+    "msw": "^0.19.0",
     "npm-run-all": "^4.0.0",
     "prettier": "^2.0.0",
     "rimraf": "^3.0.0",
