+++ src/graphql/data-sources/serlo.ts
@@ -443,17 +443,26 @@ export class SerloDataSource extends RESTDataSource {
       issuer: 'api.serlo.org',
     })
 
-    const data = (await (process.env.NODE_ENV === 'test'
-      ? super.get(`http://localhost:9009${path}`)
-      : super.get(
-          `http://${instance}.${process.env.SERLO_ORG_HOST}${path}`,
-          {},
-          {
-            headers: {
-              Authorization: `Serlo Service=${token}`,
-            },
-          }
-        ))) as unknown
+    // const data = (await (process.env.NODE_ENV === 'test'
+    //   ? super.get(`http://localhost:9009${path}`)
+    //   : super.get(
+    //       `http://${instance}.${process.env.SERLO_ORG_HOST}${path}`,
+    //       {},
+    //       {
+    //         headers: {
+    //           Authorization: `Serlo Service=${token}`,
+    //         },
+    //       }
+    //     ))) as unknown
+    const data = (await super.get(
+      `http://${instance}.${process.env.SERLO_ORG_HOST}${path}`,
+      {},
+      {
+        headers: {
+          Authorization: `Serlo Service=${token}`,
+        },
+      }
+    )) as unknown
     return await (this[setter] as (data: unknown) => Promise<T>)(data)
   }
